+++ perm/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '2.4.1'
+__version__ = '2.5.0'
 
 # Default app config for Django 1.7+
 # See https://docs.djangoproject.com/en/dev/ref/applications/#configuring-applications